+++ R/utils_default_stats_formats_labels.R
@@ -77,6 +77,7 @@ get_stats <- function(method_groups = "analyze_vars_numeric", stats_in = NULL, a
         "range", "min", "max", "median_range", "cv", "geom_mean", "geom_mean_ci",
         "geom_cv"
       ),
+      "surv_time" = c("median", "median_ci", "quantiles", "range_censor", "range_event", "range"),
       stop(
         "The selected method group (", mgi, ") has no default statistical method."
       )
@@ -136,6 +137,9 @@ get_stats <- function(method_groups = "analyze_vars_numeric", stats_in = NULL, a
 #'
 #' @param formats_in (named `vector`) \cr inserted formats to replace defaults. It can be a
 #'   character vector from [formatters::list_valid_format_labels()] or a custom format function.
+#' @param method (`character`) name of statistical method group from which to draw alternative
+#'   default formats from. E.g. For `method = "surv_time"`, the default format for `range` is
+#'   `"xx.x to xx.x"` instead of the tern default, `"xx.x - xx.x"`.
 #'
 #' @return
 #' * `get_formats_from_stats()` returns a named list of formats, they being a value from
@@ -158,7 +162,7 @@ get_stats <- function(method_groups = "analyze_vars_numeric", stats_in = NULL, a
 #' @seealso [formatting_functions]
 #'
 #' @export
-get_formats_from_stats <- function(stats, formats_in = NULL) {
+get_formats_from_stats <- function(stats, formats_in = NULL, method = NULL) {
   checkmate::assert_character(stats, min.len = 1)
   # It may be a list if there is a function in the formats
   if (checkmate::test_list(formats_in, null.ok = TRUE)) {
@@ -168,14 +172,20 @@ get_formats_from_stats <- function(stats, formats_in = NULL) {
     checkmate::assert_character(formats_in, null.ok = TRUE)
   }
 
+  out <- setNames(vector("list", length = length(stats)), stats) # Returning a list is simpler
+
   # Extract global defaults
   which_fmt <- match(stats, names(tern_default_formats))
 
   # Select only needed formats from stats
-  ret <- vector("list", length = length(stats)) # Returning a list is simpler
-  ret[!is.na(which_fmt)] <- tern_default_formats[which_fmt[!is.na(which_fmt)]]
+  out[!is.na(which_fmt)] <- tern_default_formats[which_fmt[!is.na(which_fmt)]]
 
-  out <- setNames(ret, stats)
+  # Extract any method-specific formats
+  if (!is.null(method) && method %in% names(tern_default_formats)) {
+    which_mthd_fmt <- match(stats, names(tern_default_formats[[method]]))
+    # Select only needed formats from stats
+    out[!is.na(which_mthd_fmt)] <- tern_default_formats[[method]][which_mthd_fmt[!is.na(which_mthd_fmt)]]
+  }
 
   # Modify some with custom formats
   if (!is.null(formats_in)) {
@@ -195,6 +205,12 @@ get_formats_from_stats <- function(stats, formats_in = NULL) {
 #'   variable levels will be used as the defaults, and the names of the given custom values should
 #'   correspond to levels (or have format `statistic.level`) instead of statistics. Can also be
 #'   variable names if rows correspond to different variables instead of levels. Defaults to `NULL`.
+#' @param method (`character`) name of statistical method group from which to draw alternative
+#'   default formats from. E.g. For `method = "surv_time"`, the default label for `range` is
+#'   `"Range"` instead of the tern default, `"Min - Max"`.
+#' @param control (named `list`) list of control parameters to apply to automatically adjust
+#'   default labels. E.g. If control has element `conf_level` set to `0.9`, the default label for
+#'   statistic `mean_ci` will become `"Mean 90% CI"`.
 #'
 #' @return
 #' * `get_labels_from_stats()` returns a named character vector of default labels (if present
@@ -212,7 +228,7 @@ get_formats_from_stats <- function(stats, formats_in = NULL) {
 #' get_labels_from_stats(all_cnt_occ, labels_in = list("fraction" = c("Some more fractions")))
 #'
 #' @export
-get_labels_from_stats <- function(stats, labels_in = NULL, row_nms = NULL) {
+get_labels_from_stats <- function(stats, labels_in = NULL, row_nms = NULL, method = NULL, control = NULL) {
   checkmate::assert_character(stats, min.len = 1)
   checkmate::assert_character(row_nms, null.ok = TRUE)
   # It may be a list
@@ -232,12 +248,38 @@ get_labels_from_stats <- function(stats, labels_in = NULL, row_nms = NULL) {
       for (i in lvl_lbls) out[paste(stats, i, sep = ".")] <- labels_in[[i]]
     }
   } else {
+    out <- setNames(vector("character", length = length(stats)), stats) # it needs to be a character vector
+
+    # Extract global defaults
     which_lbl <- match(stats, names(tern_default_labels))
 
-    ret <- vector("character", length = length(stats)) # it needs to be a character vector
-    ret[!is.na(which_lbl)] <- tern_default_labels[which_lbl[!is.na(which_lbl)]]
+    # Select only needed formats from stats
+    out[!is.na(which_lbl)] <- tern_default_labels[which_lbl[!is.na(which_lbl)]]
 
-    out <- setNames(ret, stats)
+    # Extract any method-specific labels
+    if (!is.null(method) && method %in% names(tern_default_labels)) {
+      which_mthd_lbl <- match(stats, names(tern_default_labels[[method]]))
+      # Select only needed formats from stats
+      out[!is.na(which_mthd_lbl)] <- tern_default_labels[[method]][which_mthd_lbl[!is.na(which_mthd_lbl)]]
+    }
+  }
+
+  # Change defaults for labels with control specs
+  if (!is.null(control)) {
+    if ("conf_level" %in% names(control)) {
+      out <- lapply(out, gsub, pattern = "[0-9]+% CI", replacement = f_conf_level(control[["conf_level"]]))
+    }
+    if ("quantiles" %in% names(control)) {
+      out[["quantiles"]] <- gsub(
+        "[0-9]+% and [0-9]+", paste0(control[["quantiles"]][1] * 100, "% and ", control[["quantiles"]][2] * 100, ""),
+        out[["quantiles"]]
+      )
+    }
+    if ("mean_pval" %in% names(control)) {
+      out[["mean_pval"]] <- gsub(
+        "p-value \\(H0: mean = [0-9\\.]+\\)", f_pval(control[["conf_level"]]), out[["mean_pval"]]
+      )
+    }
   }
 
   # Modify some with custom labels
@@ -247,7 +289,7 @@ get_labels_from_stats <- function(stats, labels_in = NULL, row_nms = NULL) {
     out[common_names] <- labels_in[common_names]
   }
 
-  out
+  unlist(out)
 }
 
 #' @describeIn default_stats_formats_labels Format indent modifiers for a given vector/list of statistics.
@@ -311,7 +353,7 @@ get_indents_from_stats <- function(stats, indents_in = NULL, row_nms = NULL) {
 #' * `tern_default_formats` is a list of available formats, named after their relevant
 #'   statistic.
 #' @export
-tern_default_formats <- c(
+tern_default_formats <- list(
   fraction = format_fraction_fixed_dp,
   unique = format_count_fraction_fixed_dp,
   nonunique = "xx",
@@ -345,7 +387,14 @@ tern_default_formats <- c(
   geom_mean_ci = "(xx.xx, xx.xx)",
   geom_cv = "xx.x",
   pval = "x.xxxx | (<0.0001)",
-  pval_counts = "x.xxxx | (<0.0001)"
+  pval_counts = "x.xxxx | (<0.0001)",
+  surv_time = list(
+    median_ci = "(xx.x, xx.x)",
+    quantiles = "xx.x, xx.x",
+    range_censor = "xx.x to xx.x",
+    range_event = "xx.x to xx.x",
+    range = "xx.x to xx.x"
+  )
 )
 
 #' @describeIn default_stats_formats_labels `character` vector that contains default labels
@@ -354,7 +403,7 @@ tern_default_formats <- c(
 #' * `tern_default_labels` is a character vector of available labels, named after their relevant
 #'   statistic.
 #' @export
-tern_default_labels <- c(
+tern_default_labels <- list(
   # list of labels -> sorted? xxx it should be not relevant due to match
   fraction = "fraction",
   unique = "Number of patients with at least one event",
@@ -388,7 +437,13 @@ tern_default_labels <- c(
   geom_mean_ci = "Geometric Mean 95% CI",
   geom_cv = "CV % Geometric Mean",
   pval = "p-value (t-test)", # Default for numeric
-  pval_counts = "p-value (chi-squared test)" # Default for counts
+  pval_counts = "p-value (chi-squared test)", # Default for counts
+  surv_time = list(
+    median_ci = "95% CI",
+    range = "Range",
+    range_censor = "Range (censored)",
+    range_event = "Range (event)"
+  )
 )
 
 # To deprecate ---------